basic electron app structure setup
