+++ gui/app.js
@@ -1,0 +1,3 @@
+'use strict';
+
+const Vue = require('vue');
