--- motd
@@ -1,33 +1,0 @@
-                                                                                                
-                                                                                                
-                                                                                                
-              0@@@8      000                                                                    
-              @@0    088@@@@8                                                                   
-             0@@80088@@@80@@@                                                                   
-              0@@@@@@80   0@@                                                                   
-                0000    008@@800            00000000000000000                000000000          
-                    008@@@@@@@@@@8          @@@88@@@@@@@@@@@@@880        08@@@@@@@@@@@@880      
-      80          08@@@@@@@@@@@@@@@8        888808@@@@@@@@@@@@@@@80   08@@@@@@@@@@@@@@@@@@0     
-     0@@80      0@@@@@@@@@@@@@@@@@@@80     08@@@0@@@888888@@@@@@@@@0 8@@@@@@@@@@@@@@@@@@8       
-     0@@@@80  08@@@@@@@@@@@@@@@@@@@@@@0 08@@@@@@0@@8       8@@@@@@@80@@@@@@@8800   00880        
-     0@@@@@@@8@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@0@@8        @@@@@@@@0@@@@@@0                    
-      8@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@888@@8       0@@@@@@@80@@@@@0                     
-       08880@@@@@@@@@@@@@@@@@@@@@@@@@@@@@000888@@@@@8888888@@@@@@@@08@@@@8                      
-           0@@@@80@@@@@@@@@@@@@@@@@@88@@@8  @@@@@@@@@@@@@@@@@@@@@808@@@@@0                      
-           8@@@0  @@@@@@@@@@@@@@@@@@0 0@@@  @@@@@@@@@@@@@@@@@@@8  8@@@@@@8                      
-           8@@8  0@@@@@@@@@@@@@@@@@@0  0@@  @@@@@@@@8888@@@@@@@8  0@@@@@@@8                     
-           0@@0  0@@80@@@@@@@@@@80@@0   @@  @@@@@@@8    8@@@@@@@8  8@@@@@@@@80       080        
-            8@0   @8 0@@@@@@@@@@8 0@0  0@8  @@@@@@@8     8@@@@@@@8  8@@@@@@@@@@88888@@@@0       
-             0@@88@8 0@@@@@@@@@@8 0@888@8   @@@@@@@8      8@@@@@@@@  08@@@@@@@@@@@@@@@@@@8      
-               08@@@@@@@@@@@@@@@@@@@@@80    @@@@@@@8       8@@@@@@@@0  088@@@@@@@@@@@@@@@80     
-                  00888@@@@@@@@@@8880       08888880        088888880      08888@@@88800        
-                                                                                                
-                                                                                                
-                                **Distributed Anonymous Object Storage**                        
-                                                  ----                                          
-                                          https://orc.network                                   
-                                                                                                
-                                 N©! 2017, Counterpoint Hackerspace Ltd                         
-                                                                                                
-                                                                                                
-
