+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "bin": {
     "orc": "bin/orc.js",
+    "orcd": "bin/orcd.js",
     "orctool": "bin/orctool.js"
   },
   "directories": {
@@ -21,7 +22,10 @@
     "coverage": "istanbul cover _mocha -- --recursive",
     "linter": "eslint ./index.js ./lib ./test",
     "generate-docs": "mkdir -p ./jsdoc && rm -r ./jsdoc && jsdoc index.js lib -r -R README.md -u ./doc -c .jsdoc.json --verbose -d ./jsdoc",
-    "publish-docs": "gh-pages -d jsdoc --repo git@github.com:orcproject/orc.git"
+    "publish-docs": "gh-pages -d jsdoc --repo git@github.com:orcproject/orc.git",
+    "start": "electron-forge start",
+    "package": "electron-forge package",
+    "make": "electron-forge make"
   },
   "keywords": [
     "orc",
@@ -56,6 +60,7 @@
     "connect": "^3.6.1",
     "cors": "^2.8.4",
     "du": "^0.1.0",
+    "electron-compile": "^6.4.1",
     "express": "^4.15.4",
     "finalhandler": "^1.0.3",
     "granax": "^2.0.7",
@@ -82,11 +87,14 @@
     "semver": "^5.1.0",
     "serve-static": "^1.12.3",
     "stream-slice": "^0.1.2",
+    "vue": "^2.4.2",
     "ws": "^3.0.0"
   },
   "devDependencies": {
     "chai": "^2.2.0",
     "coveralls": "^2.11.2",
+    "electron-forge": "^3.2.0",
+    "electron-prebuilt-compile": "1.7.5",
     "eslint": "^3.14.0",
     "form-data": "^2.2.0",
     "gh-pages": "^0.9.0",
@@ -103,5 +111,34 @@
   "optionalDependencies": {
     "bufferutil": "^3.0.2",
     "utf-8-validate": "^3.0.3"
+  },
+  "config": {
+    "forge": {
+      "make_targets": {
+        "win32": [
+          "squirrel"
+        ],
+        "darwin": [
+          "zip"
+        ],
+        "linux": [
+          "deb",
+          "rpm"
+        ]
+      },
+      "electronPackagerConfig": {},
+      "electronWinstallerConfig": {
+        "name": ""
+      },
+      "electronInstallerDebian": {},
+      "electronInstallerRedhat": {},
+      "github_repository": {
+        "owner": "",
+        "name": ""
+      },
+      "windowsStoreConfig": {
+        "packageName": ""
+      }
+    }
   }
 }
